// src/app/Menu/menu-grid/models/main-menu.ts
export interface MainMenu {
    
    Id: number;
    
    Color: string;
    
    ImageURL: string;
    
    Command: string;
    
    Caption: string;
    
    Tooltip: string;
    
    Checked: boolean;
    
    Default: boolean;
    
    Order: number;
    
    Mode: string;
    
    ObjectId: number;
    
    MenuItemGroup: number;
}